# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/templates": {
      "get": {
        "operationId": "TemplatesController_findAll",
        "summary": "Lista template",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": ["draft", "published"],
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "example": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["templates"]
      },
      "post": {
        "operationId": "TemplatesController_create",
        "summary": "Crea template (JSON)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "title": "CreateTemplateDto",
                "required": ["name", "content"],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Template Contratto"
                  },
                  "description": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string",
                    "example": "# {{titolo}}\n\nCliente: {{cliente}}"
                  },
                  "fields": {
                    "type": "array",
                    "items": {
                      "title": "TemplateFieldDto",
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string",
                          "example": "titolo"
                        },
                        "label": {
                          "type": "string",
                          "example": "Titolo"
                        },
                        "type": {
                          "type": "string",
                          "enum": ["text", "number", "date", "boolean"],
                          "example": "text"
                        },
                        "required": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["templates"]
      }
    },
    "/api/templates/{id}": {
      "get": {
        "operationId": "TemplatesController_findOne",
        "summary": "Dettaglio template",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID template",
            "schema": {}
          }
        ],
        "responses": {
          "404": {
            "description": "Template non trovato"
          }
        },
        "tags": ["templates"]
      },
      "put": {
        "operationId": "TemplatesController_update",
        "summary": "Aggiorna template (crea nuova versione)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID template",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "title": "UpdateTemplateDto",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "fields": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["templates"]
      },
      "delete": {
        "operationId": "TemplatesController_remove",
        "summary": "Elimina template",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID template",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["templates"]
      }
    },
    "/api/templates/{id}/versions": {
      "get": {
        "operationId": "TemplatesController_getVersions",
        "summary": "Cronologia versioni template",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID template",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["templates"]
      }
    },
    "/api/templates/{id}/versions/{version}": {
      "get": {
        "operationId": "TemplatesController_getVersionContent",
        "summary": "Contenuto di una versione specifica",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID template",
            "schema": {}
          },
          {
            "name": "version",
            "required": true,
            "in": "path",
            "description": "Numero versione (intero positivo)",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["templates"]
      }
    },
    "/api/templates/{id}/audit": {
      "get": {
        "operationId": "TemplatesController_getAudit",
        "summary": "Audit log del template",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID template",
            "schema": {}
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "example": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["templates"]
      }
    },
    "/api/templates/upload": {
      "post": {
        "operationId": "TemplatesController_importFile",
        "summary": "Importa template da file .md (multipart)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "title": "ImportTemplateFileDto",
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Nome del template (opzionale, default: nome file)"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["templates"]
      }
    },
    "/api/templates/validate-md": {
      "post": {
        "operationId": "TemplatesController_validateMd",
        "summary": "Valida contenuto Markdown senza creare il template",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "title": "ValidateMarkdownDto",
                "required": ["content"],
                "properties": {
                  "content": {
                    "type": "string",
                    "example": "# {{titolo}}\n\nCliente: {{cliente}}"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["templates"]
      }
    },
    "/api/templates/{id}/restore/{version}": {
      "post": {
        "operationId": "TemplatesController_restore",
        "summary": "Ripristina una versione precedente del template",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID template",
            "schema": {}
          },
          {
            "name": "version",
            "required": true,
            "in": "path",
            "description": "Versione da ripristinare",
            "schema": {}
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["templates"]
      }
    },
    "/api/templates/{id}/export": {
      "get": {
        "operationId": "TemplatesController_exportMd",
        "summary": "Scarica il template come file .md",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID template",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "File .md in download",
            "content": {
              "text/markdown": {}
            }
          }
        },
        "tags": ["templates"]
      }
    },
...
  },
  "info": {
    "title": "MAC Documents API",
    "description": "API per la generazione documentale basata su template Markdown.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
...
}